requirements.txt
# TruthLens AI - Project Requirements

# --- System Requirements (Install manually) ---
# Node.js >= 18.0.0
# npm >= 9.0.0
# ffmpeg >= 4.0.0 (Required for video processing)
# Python >= 3.9.0 (Required for Tier 3/4 local forensics)

# --- Python Dependencies (For Local Forensic Tiers) ---
# Install using: pip install -r requirements.txt
transformers>=4.38.0
torch>=2.2.0
pillow>=10.2.0
opencv-python>=4.9.0
numpy>=1.26.0
flask>=3.0.0

# --- Node.js Dependencies (Managed via npm) ---
# Managed via /frontend/package.json and /backend/package.json
# express
# cors
# dotenv
# firebase-admin
# @google/generative-ai
# nodemailer
# multer
# fluent-ffmpeg
# uuid
# react
# framer-motion
# recharts
# lucide-react
# tailwindcss
# axios
